refactor(dla): cleanup naming.
